requirements.txt
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-framework-social-oauth2==1.1.0
drf-extensions==0.7.0
gunicorn==20.0.4
ipython==7.22.0
psycopg2-binary==2.8.6
vk-api==11.9.0